Make use of DBUS_INTERFACES qmake variable
This commit eliminates the need to store (in git) the generated D-Bus
interface files and makes make-clean work for them.

The problem with including commondbustypes.h is solved by including not only
the generated .h file but also generated .cpp file as well as its moc_* file
inside the associated API class' source file after commondbustypes.h is
included.  Note that the .h file still needs to be explicitly included -
without being referenced, qmake would not detect any dependency to it and it
would not get generated.  See networkmanager.cpp:

 #include "commondbustypes.h"
 #include "connman_manager_interface.h"
 #include "connman_manager_interface.cpp" // not bug
 #include "moc_connman_manager_interface.cpp" // not bug