Use shadow build to keep source tree clean
The main reason to do this is for code coverage analyzis to work. The tool
`lcov' is happy with shadow build but it does not work with OBJECTS_DIR,
MOC_DIR and similar qmake variables set in order to keep source tre clean of
build artifacts.

 $ cd build/
 $ qmake ..
 $ make

To clean simply run `rm -rf build/*' from project root.