[connman-qt] Add properties to selectively enable services and technology updates.
Updating services and technologies is unnecessary for some use cases
and states. Such as when an application is only interested in the
offlineMode or state properties or when the functionality is not being
actively used by the user.

Two properties are added servicesEnabled and technologiesEnabled, both
default to true. When servicesEnabled is false NetworkManager will
report that there are no network services available. When
technologiesEnabled is false NetworkManager will report that there are
no technologies available.

In many applications a single global shared instance of NetworkManager
is created. When using these properties applications should ensure that
the values are appropriate for all users of the NetworkManager
instance.