[connman-qt] Don't show hidden services that are really connected.
When connecting to a hidden service, the real known service that gets
connected is created, so essentially, there are two entries.

Use the new hidden and bssid properties to best guess and hide the
hidden service with no name property.

This will not work correctly if there is more than one hidden service
per bssid, in which case all hidden AP's will get hidden.