[connman-qt] Remove debug message about unknown technology
This message can be misleading as it is frequently displayed
referencing known technologies. Due to the use of asynchronous DBus
calls the NetworkManager::getTechnology() function may be called prior
to the available technology list being populated. This happens, for
example, when the name property of the TechnologyModel is set during
construction.