[libconnman-qt] Add API for accessing connman VPNs. Contributes to JB#45378
Connman provides the vpn-manager and vpn-connection dbus interfaces for
managing VPN configurations. These changes provide a Qt API wrapper for
these interfaces, along with a model exending QAbstractListModel of the
available VPN configurations.

Much of the API is modeled on the VpnModel that can be found in
nemo-qml-plugin-systemsettings, stripped back to remove
Settings-specific functionality.