[libconnman-qt] Set multiple VPN properties with one D-Bus call. JB#46120
Changed to use SetProperties implemented to ConnMan vpnd when modifying
the properties of a VPN connection. SetProperties accepts an a{sv} so
the marshalled QVariantMap can be used.

This improves the performance as only one D-Bus call is to be made when
a VPN properties are saved. Downside is that error on individual
property cannot be received. InvalidProperty is received only when all
properties are invalid.