[libconnman-qt] Use SetProperty for multiple properties. JB#46120
ConnMan upstream requested to have all property setting functionality
behind SetProperty, and to drop SetProperties from the proposed patch.
This change accomodates that change.

SetProperty can accept DBUS_TYPE_ARRAY as variant. If the name is other
than "UserRoutes" it will be parsed as an array of dict, as it would
have been done with the removed SetProperties D-Bus method. The
QVariantMap is wrapped inside a QVariant, and the string for message is
set as "properties" to have the same value as in ConnMan vpn-provider.c
D-Bus method registration, to accommodate also the potential future use
of having requirement for the dict to be named as "properties".