[libconnman-qt] Support DefaultRoute VPN property. JB#48797
Add support for setting default route property for VPN connections.
ConnMan has this support already, there wasn't any UI or library support
for it.

ConnMan has the "DefaultRoute" option as string with values
"true"|"false". Added conversion of DefaultRoute variant string to bool
when receiving it from ConnMan and doing the vice versa at sending
the content to ConnMan.

Publish the value as boolean for components using the library. Store
the value in m_defaultRoute if the setters and getters are used, which
are also implemented.