[libconnman-qt] Fix ethernet service visibility in UI. Fixes JB#56522
Do not compare the list counts with ethernet services. This may result
in a situation where the available services are not shown when the
ethernet services are to be requested on a newly flashed device or when
there hasn't been many other services saved or in use. The ethernet
services work in a bit different way as they on the first run are not
saved at all but are created by ConnMan when LOWER_UP is being detected
and thus, they are shown as available. This makes the available and
saved ethernet services to be shown always whenever there is at least
one of either.