[libconnman-qt] Follow DefaultRoute signal and use ordered list. JB#58240
ConnMan sends additional PropertyChanged signal that indicates to which
defaultroute has been changed to. Use this to see changes in the
service, which includes changes to VPN being the default route.

Improve default route management by using ordered list to check for
default route. If the topmost service is connected it is used as a
default route because ConnMan reports the list of services in ordered
manner and the top one is the default route. No need to parse routing
tables for the default route.

This change accounts for VPNs being as default and the next one in the
list is ordered to be the transport service used by the VPN. As VPN
services are not included in networkmanager use the next connected as
the default and keep network in connected state.