[libconnman-qt] Track service cache changes for def route update. Fixes JB#61555
Do the default route update when there are changes in the service cache.
This prevents doing unnecessary list traversals when there are no
updates or when the default route is kept as the same. Also drop out
from the services cache traversal when first not connected service is
encountered since the service states are used for the ordered list
originating from ConnMan.

And cleanup some old debug to make output more clear.

Drop excess call to updateDefaultRoute() - called in updateServices().