[libconnman-qt] Share NetworkManager instances for QML API. JB#61563
The NetworkManager is mostly just a D-Bus wrapper. The target is just
one instance in Connman so we should be able to also share the D-Bus proxy
instance here. Though just doing it now for QML. We know we don't need
to care about threading for that side, and in practice that's where
this is mostly used.

Some minor differences on API:
- createSession and destroySession not anymore included. Former wasn't
usable in QML really as the returned QDBusObjectPath should have been
converted to string.
- setOfflineMode() and setSessionMode() were slots, thus public in QML,
but they are redundant when there's already a property that can be
assigned. These are still kept but they warn about being deprecated.
- Some value parameters removed from property change signals.
These were more for C++ API, whereas QML code commonly just checks the
property value. Assuming minor enough to allow change.