[libconnman-qt] Deprecate NetworkManagerFactory with reference counted api. JB#61563
Instances created with NetworkManagerFactory or NetworkManager::instance()
not only leaked memory (as in not relesed at exit), but also made the
process track Connman changes even if the need was short time.

So instead providing now QSharedPointer wrapped instance with
NetworkManager::sharedInstance().

This is a small ABI break for the classes having NetworkManager pointer
in public class's 'private:' part.