[libconnman-qt] Support tethering client list signals and method. JB#60565
Add support for the tethering client list from ConnMan. The content is a
(dict) QVariantMap with MAC address as key, containing a (dict)
QVariantMap with the client specifics (IP, technology and address type).
This is parsed to be a QVariantList with each entry being:
"mac", "address", "addresstype" and "technology". "Version" is not used
yet as it is not used in ConnMan.

Add listening of the signal and notify new and disconnected tethering
clients to UI when there is a change by first retrieving the new list of
tethering clients. Clear the list when tethering is disabled and notify
upwards.

+ add new copyright lines for 2025.