[libconnman-qt] Add support for mDNS. JB#63356
The value "mDNS" is the real value that is set, and loaded for the
service. This is signaled always when the value changes. The
"mDNS.Configuration" is the requested configuration value for mDNS set
via D-Bus. This is not always communicated upwards when "mDNS" changes,
but the "mDNS" value is changed based on changes made with
"mDNS.Configuration", and "mDNS" is not read from D-Bus SetProperty()
call content. Thus, "mDNS" is the effective value and setting happens
with "mDNS.Configuration".